move error helper to util directory
